Split up 1010CG e2e tests
